auth is working with frontend, products can now be edited, deleted and added from the frontend

+++ pizzeria-backend/Controllers/ProductsController.cs
@@ -13,7 +13,6 @@ namespace pizzeria_backend.Controllers
         private readonly IProductService _productService = productService;
         private readonly IAzureBlobStorageService _azureBlobStorageService = azureBlobStorageService;
 
-
         [HttpPost()]
         [Authorize]
         [Produces("application/json")]
@@ -35,11 +34,21 @@ namespace pizzeria_backend.Controllers
         public async Task<IActionResult> GetProduct(int id)
         {
             var pr = await _productService.GetProduct(id);
+
             if (pr == null)
             {
                 return NotFound("Product not found");
             }
 
+            return Ok(pr);
+        }
+
+        [HttpGet("")]
+        [Authorize(Policy = "Admin")]
+        [Produces("application/json")]
+        public async Task<IActionResult> GetAllProduct()
+        {
+            var pr = await _productService.GetAllProducts();
 
             return Ok(pr);
         }
@@ -53,24 +62,29 @@ namespace pizzeria_backend.Controllers
             {
                 return NotFound("There are no items in the menu");
             }
+
             return Ok(menu);
         }
 
         [HttpDelete("{id}")]
-        [Authorize]
+        [Authorize(Policy = "Admin")]
         [Produces("application/json")]
         public async Task<IActionResult> DeleteProduct(int id)
         {
             var pr = await _productService.DeleteProduct(id);
+
             if (pr == null)
             {
                 return NotFound("Product not found");
             }
+
+            await _azureBlobStorageService.DeleteBlobAsync(pr.Image.Split("/").Last());
+
             return Ok(pr);
         }
 
         [HttpPatch("{id}")]
-        [Authorize]
+        [Authorize(Policy = "Admin")]
         [Produces("application/json")]
         public async Task<IActionResult> UpdateProduct(int id, [FromForm] ProductDto product)
         {
@@ -80,10 +94,19 @@ namespace pizzeria_backend.Controllers
                 return NotFound("Product not found");
             }
 
+            if (product.Image != null)
+            {
+                await _azureBlobStorageService.DeleteBlobAsync(pr.Image.Split("/").Last());
+                var image = (await _azureBlobStorageService.UploadBlobAsync(product.Image.OpenReadStream(), product.Image.FileName)).ToString();
+                pr = ConvertToProduct(product, image, id);
 
-            await _azureBlobStorageService.DeleteBlobAsync(pr.Image.Split("/").Last());
-            var image = (await _azureBlobStorageService.UploadBlobAsync(product.Image.OpenReadStream(), product.Image.FileName)).ToString();
-            pr = ConvertToProduct(product, image, id);
+            }
+            else
+            {
+                pr = ConvertToProduct(product, pr.Image, id);
+
+            }
+
             pr = await _productService.UpdateProduct(pr);
 
             return Ok(pr);
@@ -103,13 +126,28 @@ namespace pizzeria_backend.Controllers
             };
         }
 
-
         private static Product ConvertToProduct(ProductDto productDto, string image, int id)
         {
+
             return new Product
             {
                 Id = id,
                 Image = image,
+                Name = productDto.Name,
+                Description = productDto.Description,
+                CategoryId = productDto.CategoryId,
+                Price = productDto.Price,
+                IsAvailable = productDto.IsAvailable,
+                IsInMenu = productDto.IsInMenu,
+            };
+        }
+
+        private static Product ConvertToProduct(ProductDto productDto, int id)
+        {
+
+            return new Product
+            {
+                Id = id,
                 Name = productDto.Name,
                 Description = productDto.Description,
                 CategoryId = productDto.CategoryId,
